requirements.txt
click==7.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pep8==1.7.1
Flask==1.0
gunicorn==19.7.1
cloudant==2.4.0
flasgger==0.6.4
Flask-Cors
Flask-RESTful==0.3.6
flask-restplus
setuptools>=18.0
cython
Werkzeug==0.16.1
numpy==1.19.1
pandas==1.1.1
sqlalchemy==1.3.19
pymysql==0.9.2
python-dotenv
APScheduler==3.6.3
websocket-client
kafka-python==2.0.2
tweepy